requirements.txt
# Knjižnica API v8.6.0 — requirements.txt
# Render.com free tier (Python 3.11)
# ─────────────────────────────────────────────────────────────────────────────

# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
aiosqlite>=0.20.0

# Auth & Security
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt==4.0.1             # pinana verzija — novije verzije mijenjaju API

# Validation
pydantic>=2.0.0
pydantic[email]>=2.0.0

# HTTP klijent (za interno testiranje)
httpx>=0.27.0

# Async
aiofiles>=23.0.0

# WebSocket
websockets>=12.0

# Env & Utils
python-dotenv>=1.0.0

# ─────────────────────────────────────────────────────────────────────────────
# UKLONJENO (nije potrebno za Render free tier):
#   celery, redis, flower   — task queue (nije korišten)
#   psycopg2-binary         — PostgreSQL driver (koristimo SQLite)
#   pytest, black, isort, mypy, types-python-jose  — dev alati
# ─────────────────────────────────────────────────────────────────────────────
